useFetch를 통한 통신, util/api 디랙토리 분리
