+++ App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
-import { AuthContextProvider } from './src/api/Auth';
+import { AuthContextProvider } from './src/util/Auth';
 
 import Navigator from "./src/navigation";
 import { useFonts } from 'expo-font';
