+++ src/api/campagin.ts
@@ -1,0 +1,9 @@
+import { ip } from "./ip"
+import useFetch from "./useFetch"
+
+export const addCampagin = () => {
+    return useFetch(`${ip}//campaign/register`, {
+        method: "POST",
+        
+    })
+}
