+++ src/api/index.ts
@@ -1,0 +1,5 @@
+import useFetch from "./useFetch"
+
+export const getRandomCat = () => {
+    return useFetch("http://aws.random.cat/meow");
+}
