--- src/axios/campagin.ts
@@ -1,8 +1,0 @@
-import axios from "axios"
-import { ip } from "./ip"
-
-export const addCampagin = () => {
-    return axios.post(`${ip}/campagin`, {
-
-    })
-}
