+++ src/components/Campagin/MakeCouponModal.tsx
@@ -4,10 +4,10 @@ import { Coupon, ModalStackParamList } from '@types';
 import { ImgPicker, OutLineButton, InputModal } from '../../atoms';
 import { ScrollWrapper, SubTitle } from '../../atoms/styled'
 import DateTimePicker from '@react-native-community/datetimepicker';
-import perventGoBack from '../../api/perventGoBack';
+import perventGoBack from '../../util/perventGoBack';
 import { campaginNavigation } from '../../navigation/useNavigation';
 import { RouteProp, useNavigation, useRoute } from '@react-navigation/core';
-import { isEditCoupon } from '../../api/util';
+import { isEditCoupon } from '../../util';
 
 const MakeCouponModal = () => {
     const campaginNav = campaginNavigation();
