+++ src/components/GamePlay/index.tsx
@@ -1,7 +1,10 @@
 import { useNavigation } from '@react-navigation/core'
 import React from 'react'
 import { Button, Text } from 'react-native'
+import { Image } from 'react-native-elements'
 import styled from 'styled-components/native'
+import { getRandomCat } from '../../api'
+import useFetch from '../../api/useFetch'
 
 interface Props {
 
@@ -9,16 +12,25 @@ interface Props {
 
 export default (props: Props) => {
     const navigation = useNavigation();
+    const { data, err, loading, refetch } = getRandomCat();
+
     return (
         <Container>
             <Text>Map</Text>
             <Button
                 title="play"
                 onPress={() => navigation.navigate("Game")} />
+            <Button
+                title="RANDOM"
+                onPress={refetch} />
+            <Image source={{ uri: data ? data.file : null }}
+                style={{ width: 200, height: 200 }} />
+            <Text>{loading ? "loading" : " "}</Text>
+            <Text> {err} </Text>
+
         </Container>
     )
 }
-
 
 const Container = styled.View`
     flex: 1;
